refactor: extract session runtime helper functions
